Fix font names in theme.txt file so they match names in .pf2 files.
Opening the Hack-18.pf2 file in a hex editor shows that the NAME set
for this font is 'Hack Regular 18', and not 'Hack 18'. Similarly, all
the other Hack family fonts in the theme directory have names
beginning with 'Hack Regular...'.

Consequently, in the theme.txt file when the title-font, message-font
and terminal-font are set to 'Hack 18', and the item_font and
selected_item_font are set to 'Hack 22', GRUB cannot find the fonts
and ends up displaying other fonts, which on my system are HUGE--so
big, in fact, that some of the menu entries do not fit on the screen.

With these corrections made to the font names in the theme.txt file,
GRUB finds the fonts in the theme directory, and the theme displays
properly, with correct-sized fonts.